requirements.txt
phidata
python-dotenv
groq
sqlalchemy
fastapi #for setuping playground
python-multipart
uvicorn